% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{bw}
\alias{bw}
\title{Density kernel bandwidth}
\usage{
bw(x)
}
\arguments{
\item{x}{data, as a matrix. Each row is a data points, each column is a dimension. NA values will be ignored.}
}
\value{
a vector of bandwidths, the length is the number of columns in x.
}
\description{
The Silverman's rule of thumb kernel bandwidth for Gaussian 2nd order density kernel, for multidimensional data.
}
