% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Performance Metrics}
\usage{
calculate_metrics(scores, outliers, contamination)
}
\description{
Calculate various metrics to assess outlier detection quality.
}
\keyword{internal}
