% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_filter_outlier.R
\name{control_filter_outlier}
\alias{control_filter_outlier}
\title{Control_object}
\usage{
control_filter_outlier(
  numeric_method = "mean_sd",
  discrete_method = "prop",
  threshold = NULL,
  conf_int = NULL,
  prop = NULL,
  n = NULL,
  freq = NULL,
  min_times = NULL,
  ties_method = "min",
  na_action = "keep"
)
}
\arguments{
\item{numeric_method}{dd}

\item{discrete_method}{m}

\item{threshold}{for some numeric methods}

\item{conf_int}{conf int for t-test}

\item{prop}{propotion of factors/char (0 - 1)}

\item{ties_method}{factors}

\item{na_action}{"keep" NA is the default}
}
\value{
control_object
}
\description{
Control_object
}
\examples{
control_filter_outlier()
}
