% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_outlier.R
\name{extract_outlier}
\alias{extract_outlier}
\title{extract outliers from}
\usage{
extract_outlier(x, ..., .ptype = logical())
}
\arguments{
\item{x}{outlier object}

\item{...}{not used}

\item{.ptype}{type to}
}
\value{
vector or tibble
}
\description{
extract outliers from
}
\examples{
print("hello")
}
