% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{flag}
\alias{flag}
\title{Flag outliers}
\usage{
flag(x, thresh = 1.5, nmax = NULL, side = NULL, crit = "lof",
  asInt = TRUE, k = 5, metric = "euclidean", q = 3)
}
\arguments{
\item{x}{a matrix, data frame or vector of data points (a vector will be understood as 1D data, equivalent to a 1-column matrix). Each row is a data point and each column is a dimension. NA values are allowed and will produce NAs in the output.}

\item{thresh}{threshold for the outlier factor above which the data point is flagged as an outlier. The outlier factor is the LOF for the LOF method and the "1+odds of outlying" for the Chauvenet method}

\item{nmax}{the maximum number of outliers to remove. If NULL, ignored.}

\item{side}{if set to 'left', 'right' or 'both' (can be abreviated to one letter and case-insensitive) will flag only the outliers on the left, right or both ends of the 1D distribution. If NULL, all outliers will be flagged. If the data is not 1D, \code{side} will be ignored. Note that for the methods that only find outliers on the sides of the distribution (e.g Chauvenet) NULL and 'both' give equivalent results.}

\item{crit}{criterion to use for identifying outliers. Currently, can be either 'lof' or 'chauvenet'. Can be abbreviated to 3 first letters, case-insentitive}

\item{asInt}{if TRUE, the flag values will be integers (1 for outlier and 0 otherwise). If FALSE, boolean}

\item{k}{number of nearest neighbors for the LOF calculation}

\item{metric}{distance metric to use. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}

\item{q}{the power of the Minkowski distance.}
}
\value{
a boolean or integer (depending on \code{asInt}) vector of the same length as the number of points in the data, containing 1 (TRUE) if a data point is an outlier, 0 (FALSE) if it is not and NA if a point in the data contained NA value(s), the
}
\description{
Flag outliers in a multidimensional data set
}
