% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_n_cores}
\alias{get_n_cores}
\title{Get Number of Cores for Parallel Processing}
\usage{
get_n_cores(max_tasks)
}
\arguments{
\item{max_tasks}{Maximum number of parallel tasks to run}
}
\value{
Integer number of cores to use
}
\description{
Internal helper to determine number of cores to use for parallel processing
in a CRAN-compliant way. Respects R CMD check limitations.
}
\keyword{internal}
