% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers.R
\name{get_outliers}
\alias{get_outliers}
\title{Title}
\usage{
get_outliers(x, ..., .ptype = logical())
}
\arguments{
\item{x}{outlier object}

\item{...}{not used}

\item{.ptype}{description}
}
\description{
Title
}
