% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_outlier.R
\name{identify_outlier}
\alias{identify_outlier}
\title{Identity  filter}
\usage{
identify_outlier(.data, ..., control = control_outlier())
}
\arguments{
\item{.data}{data.frame or tibble}

\item{...}{variable(s) to identify outlier}

\item{control}{control_obj}
}
\value{
data.frame or tibble
}
\description{
Identity  filter
}
\examples{

mtcars["var_1"] <- c(rnorm(30), -100, 100)
mtcars["var_2"] <- c(-50, 32, rnorm(30))
filtred <-
  mtcars |> identify_outlier(var_1, var_2)

print(filtred)
# ── 4 Outliers were removed of 32 rows. ──────────────────────────────────────────────────
}
