% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{impute}
\alias{impute}
\title{Impute outliers}
\usage{
impute(x, flag = NULL, fill = "mean", thresh = 1.5, nmax = NULL,
  side = NULL, crit = "lof", k = 5, metric = "euclidean", q = 3, ...)
}
\arguments{
\item{x}{a matrix, data frame or vector of data points (a vector will be understood as 1D data, equivalent to a 1-column matrix). Each row is a data point and each column is a dimension. NA values are allowed and will produce NAs in the output.}

\item{flag}{a boolean or integer (0-or-1) vector flagging outliers, such as produced by the function \code{flag}. If NULL, further arguments will be used to compute it here by calling \code{flag}.}

\item{fill}{method for imputing (or removing) outliers. If numeric or NA, it is the value that will replace the outliers. It the data is K-dimensional, \code{fill} is expected to be a vector of length K. If longer, the first K components will be used, and if shorter, the vector will be extended by NAs.
Alternatively, \code{fill} can be a character string. Values 'mean' and 'median' replace outliers with the mean or (multidimensional) median of the rest of the remaining data,
'remove' removes the outliers by calling the function \code{purge}. Can be abbreviated to the first 3 letters, case-insensitive.}

\item{thresh}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{nmax}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{side}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{crit}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{k}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{metric}{distance metric to be used in LOF (if \code{flag} is not provided) as well as for multidimensional median if \code{fill} is 'median'. A choice of 'euclidean','maximum','manhattan','canberra','minkowski', or 'binary'. Can be abbreviated to first three letters, case-insensitive.}

\item{q}{power in Minkowski metric, used if \code{fill='median'} and \code{metric='minkowski'}}

\item{...}{passed to \code{Rcgmin} if the argument \code{fill} is 'median' and data is multidimensional}
}
\value{
object like \code{x} but with outliers imputed.
}
\description{
Impute detected outliers in a multidimensional data set
}
\details{
The output object will be a vector, a matrix or a data-frame, depending on what \code{x} was. Row names, column names or (if \code{x} was a named vector) names will be kept.
}
