% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof.R
\name{lof}
\alias{lof}
\title{Local Outlier Factor (LOF) values}
\usage{
lof(x, k = 5, metric = "euclidean", q = 3)
}
\arguments{
\item{x}{a matrix or data frame of data points. Each row is a data point, each column is a dimension.}

\item{k}{number of nearest neighbors in the LOF calculation (default 5).}

\item{metric}{distance metric to use. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}

\item{q}{the power of the Minkowski distance.}
}
\value{
a vector of LOF values, in the same order as the data points in \code{x}. NA values will be returned for those data points which contained at least one NA coordinate.
}
\description{
Calculates the local outlier factor (LOF) for multidimensional data
}
