% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlieR-package.R
\docType{package}
\name{outlieR-package}
\alias{outlieR}
\alias{outlieR-package}
\title{outlieR: Identify, Filter and Visualisation of Outliers}
\description{
Identify, Filter and Visualisation of Outliers using tidy principles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedersebastian/outlieR}
  \item \url{https://pedersebastian.github.io/outlieR/}
  \item Report bugs at \url{https://github.com/pedersebastian/outlieR/issues}
}

}
\keyword{internal}
