% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{outlier}
\alias{outlier}
\title{Outlier searching}
\usage{
outlier(input, superior = TRUE, out_type = "logical")
}
\arguments{
\item{input}{data frame}

\item{superior}{if TRUE, returns only the superior outliers}

\item{out_type}{'logical', 'absolute' or 'percentage'}
}
\value{
Data frame, type defined by out_type parameter
}
\description{
Returns a boolean data frame with the same dimensions of input dataframe
where the TRUES points to the input outliers
}
\examples{
input=data.frame(a=c(10,11,rnorm(20),14,-20),b=c(5,6,rnorm(20),7,-5))
output <- outlier(input)

}

