% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_more.R
\name{outlier_flere}
\alias{outlier_flere}
\title{Outlier filter}
\usage{
outlier_flere(
  .data,
  ...,
  method = c("mean_sd", "MAD", "IQD", "t_test"),
  threshold = "default",
  conf_int = NULL,
  na_action = c("keep", "omit")
)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{...}{variables to filter outlier}

\item{method}{method of filter outlier.}

\item{threshold}{threshold for the other methods than t-test. Default is 3 for mean and sd method and MAD method, IQD uses 2.2.}

\item{conf_int}{confidence interval if method is t-test}

\item{na_action}{to also filter NA´s ?}
}
\value{
data.frame or tibble
}
\description{
Outlier filter
}
\examples{

mtcars["V1"] <- c(rnorm(30), -100, 100)
mtcars["V2"] <- c(-50, 32, rnorm(30))
filtred <-
  mtcars |> outlier_flere(V1, V2)

paste(nrow(mtcars), "rows before filtered and", nrow(filtred), "rows after")
 #"32 rows before filtered and 28 rows after"
}
