% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot Feature Importance for Outliers}
\usage{
plot_feature_importance(x, n_top = 10)
}
\description{
Plot Feature Importance for Outliers
}
\keyword{internal}
