% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_score_distribution}
\alias{plot_score_distribution}
\title{Plot Anomaly Score Distribution}
\usage{
plot_score_distribution(x, n_top = 10)
}
\description{
Plot Anomaly Score Distribution
}
\keyword{internal}
