% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{print.outlier_detector}
\alias{print.outlier_detector}
\title{Print Method for Outlier Detector}
\usage{
\method{print}{outlier_detector}(x, ...)
}
\arguments{
\item{x}{An outlier_detector object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Outlier Detector
}
