% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{purge}
\alias{purge}
\title{Remove outliers}
\usage{
purge(x, flag = NULL, level = 0.1, nmax = NULL, side = NULL,
  crit = "lof", k = 5, metric = "euclidean", q = 3)
}
\arguments{
\item{x}{a matrix, data frame or vector of data points (a vector will be understood as 1D data, equivalent to a 1-column matrix). Each row is a data point and each column is a dimension. NA values are allowed and will produce NAs in the output.}

\item{flag}{a boolean or integer (0-or-1) vector flagging outliers, such as produced by the function \code{flag}. If NULL, further arguments will be used to compute it here by calling \code{flag}.}

\item{level}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{nmax}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{side}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{crit}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{k}{passed to the function \code{flag} if the argument 'flag' is NULL or missing}

\item{metric}{distance metric to be used in LOF (if \code{flag} is not provided). A choice of 'euclidean','maximum','manhattan','canberra','minkowski', or 'binary'. Any unambiguous substring can be given, case insensitive.}

\item{q}{power in Minkowski metric, used if \code{fill='median'} and \code{metric='minkowski'}}

\item{fill}{method for imputing (or removing) outliers:'mean' and 'median' replace outliers with the mean or (multidimensional) median of the rest of the remaining data,
'remove' removes the outliers. Any unambiguous substring can be given, case insensitive.}
}
\value{
object like \code{x} but with outliers removed.
}
\description{
Remove detected outliers in a multidimensional data set
}
\details{
The output object will be a vector, a matrix or a data-frame, depending on what \code{x} was. Row names, column names or (if \code{x} was a named vector) names will be kept.
}
