% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/tidies.R, R/tidyselect.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{glance}
\alias{augment}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\alias{last_col}
\alias{any_of}
\alias{all_of}
\alias{where}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect:all_of]{any_of}}, \code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:everything]{last_col}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{where}}}
}}

