% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rel_change}
\alias{rel_change}
\title{Relative Change}
\usage{
rel_change(x, xref)
}
\arguments{
\item{x}{a numeric or complex vector or array}

\item{xref}{ref that x is compared to, must have same size of x}
}
\value{
data of same size of x and xref
}
\description{
Relative Change calculation, as stated in \link{https://en.wikipedia.org/wiki/Relative_change_and_difference}
}
\examples{
output <- rel_change(1,0.4)
print(output)

}

