% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_cont_dates.r
\name{scale_x_cont_dates}
\alias{scale_x_cont_dates}
\title{scale_x_cont_dates}
\usage{
scale_x_cont_dates(
  ...,
  name = name,
  business.dates,
  max.major.breaks = 5,
  max.minor.breaks = max.major.breaks * 5,
  breaks = bd_breaks(business.dates)
)
}
\arguments{
\item{...}{Argumentos adicionales que se pasan a la función scale_x_continuous}

\item{name}{Nombre del eje x}

\item{business.dates}{Un vector de fechas de negocios. Acá le suelo pasar el vector de fechas de los días que tengo en el df con el que lo llamo}

\item{max.major.breaks}{Cantidad máxima de breaks mayores}

\item{max.minor.breaks}{Cantidad máxima de breaks menores}

\item{breaks}{Los breaks que se van a usar. Por defecto se usa la función `bd_breaks` de la librería `bdscale`}
}
\value{
Una escala x para fechas de negocios
}
\description{
Una función que crea una escala x para fechas de negocios. De esta manera permite que las 
barras de los ejes estén todas contiguas. Sino separa por semanas.
Tuve que hacer un override de la función scale_x_bd de la libreria `bdscale` porque hardcodearon el nombre
del eje x y no me dejaba cambiarlo.
}
