% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_plot.R
\name{summarise_outlier}
\alias{summarise_outlier}
\title{Summary of Outliers}
\usage{
summarise_outlier(object, ...)
}
\arguments{
\item{object}{object created by ´identify_outlier()´}

\item{...}{not used}
}
\value{
ggplot
}
\description{
Summary of Outliers
}
\examples{
identify_outlier(tibble::as_tibble(ggplot2::txhousing), everything()) |>
  summarise_outlier()
}
