% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_cont_dates.r
\name{t2bd}
\alias{t2bd}
\title{t2bd}
\usage{
t2bd(ts, business.dates)
}
\description{
No entiendo por qué tuve que buscar esta función y pegarla acá.
La usa scale_x_cont_dates y por alguna razón no la encuentro exportada desde ahi
Entonces la armé de vuelta para usarla
}
