% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{theme_outlier}
\alias{theme_outlier}
\title{Theme for outlier plots}
\usage{
theme_outlier(bg_colour = "#fafafa", ...)
}
\arguments{
\item{...}{dots}

\item{bg_col}{background colour}
}
\description{
Theme for outlier plots
}
