% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{tune_bayesian_optimization}
\alias{tune_bayesian_optimization}
\title{Bayesian Optimization for Hyperparameter Tuning}
\usage{
tune_bayesian_optimization(data, param_space, contamination, parallel, verbose)
}
\description{
Bayesian Optimization for Hyperparameter Tuning
}
\keyword{internal}
