% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_filter_outlier.R
\name{control_filter_outlier}
\alias{control_filter_outlier}
\title{Control_object}
\usage{
control_filter_outlier(
  numeric_method = "mean_sd",
  discrete_method = "prop",
  threshold = NULL,
  conf_int = NULL,
  prop = NULL,
  n_vars = NULL,
  freq = NULL,
  min_times = NULL,
  ties_method = "min",
  na_action = "keep",
  ...
)
}
\arguments{
\item{numeric_method}{dd}

\item{discrete_method}{m}

\item{threshold}{'NULL' for the other methods than t-test.
Default is 3 for mean and sd method and MAD method, IQD uses 2.2.}

\item{conf_int}{conf int for t-test}

\item{prop}{proportion of factors/char (0 - 1)}

\item{n_vars}{n_vars if}

\item{freq}{f}

\item{min_times}{m}

\item{ties_method}{factors}

\item{na_action}{"keep" NA is the default}

\item{...}{not currently used}
}
\value{
control_object
}
\description{
Control_object
}
\examples{
control_filter_outlier()
}
