% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_outlier.R
\name{display_outlier}
\alias{display_outlier}
\title{Display outliers of an}
\usage{
display_outlier(object, ...)
}
\arguments{
\item{object}{object created by identify_outlier()}

\item{...}{not used}
}
\value{
plot
}
\description{
Display outliers of an
}
\examples{
print("kommer")

}
