% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outlier.R
\name{filter_outlier}
\alias{filter_outlier}
\title{Filter a data frame/ tibble of outliers}
\usage{
filter_outlier(object, ...)
}
\arguments{
\item{object}{an object created by identify_outlier}

\item{...}{not currently fused}
}
\value{
a tbl/dataframe
}
\description{
Filter a data frame/ tibble of outliers
}
\examples{
print("hei")
}
