% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outlier.R
\name{filter_outlier}
\alias{filter_outlier}
\title{Outlier filter}
\usage{
filter_outlier(.data, ..., control = control_filter_outlier())
}
\arguments{
\item{.data}{data.frame or tibble}

\item{...}{variable(s) to filter outlier}

\item{control}{control_obj}
}
\value{
data.frame or tibble
}
\description{
Outlier filter
}
\examples{

mtcars["V1"] <- c(rnorm(30), -100, 100)
mtcars["V2"] <- c(-50, 32, rnorm(30))
filtred <-
  mtcars |> filter_outlier(V1, V2)

paste(nrow(mtcars), "rows before filtered and", nrow(filtred), "rows after")
# "32 rows before filtered and 28 rows after"
}
