% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_outlier}
\alias{theme_outlier}
\title{theme_outlier}
\usage{
theme_outlier(base_size = 11, base_family = "Microsoft Sans Serif")
}
\value{
objeto devuelto por theme_grey con las características de outlier
}
\description{
Define la estética de los gráficos
agrega dos variables globales .pie y .paleta para poder ser usada en la construcción de gráficos
}
